:ok_hand: Remove old functions
