+++ .github/workflows/ci.yml
@@ -36,4 +36,4 @@ jobs:
       - run: mix compile --warnings-as-errors
       - run: mix format --check-formatted
       - run: mix credo --strict
-      - run: mix test
+      - run: mix test --trace --raise
